+++ src/web/views/decision-fallback-successful.njk
@@ -11,6 +11,6 @@
     classes: "govuk-!-margin-bottom-5"
   }) }}
 
-  {% include "locales/" + language + "/decision.njk" %}
+  {% include "locales/" + language + "/decision-fallback-successful.njk" %}
 
 {% endblock %}
