+++ src/web/views/decision-fallback-unsuccessful.njk
@@ -8,6 +8,6 @@
         classes: "govuk-!-margin-bottom-5"
     }) }}
 
-    {% include "locales/" + language + "/unsuccessful/" + eligibilityStatus + ".html" %}
+    {% include "locales/" + language + "/decision-fallback-unsuccessful/" + eligibilityStatus + ".html" %}
 
 {% endblock %}
